+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # GDMC-HTTP 1.4.6 (Minecraft 1.19.2 + 1.20.2)
 
 - NEW: To enhance backwards compatibility, all features introduced in GDMC-HTTP 1.4.0 are now also available for Minecraft 1.19.2!
+- FIX: When the `doBlockUpdates=false` parameter is set on `PUT /blocks` discard placement instructions for duplicate block positions except the last one for that position. This prevents an undefined behaviour that made it impossible to predict in what order placement instructions targeting the same block position would resolve.
 
 # GDMC-HTTP 1.4.5 (Minecraft 1.20.2)